Move the define_css_keyword_enum macro to the style crate.
